arreglo de la seguridad y creación de vista para 'acceso denegado'

+++ proyecto/src/main/java/com/Proyecto/controller/AdminController.java
@@ -19,12 +19,12 @@ import org.springframework.web.bind.annotation.PathVariable;
 @Controller
 @RequestMapping("/admin")
 public class AdminController {
-@Autowired
-private LibrosService librosService;
-@Autowired
-private UsuarioService usuarioService;
-@Autowired
-private CompraService compraService;
+    @Autowired
+    private LibrosService librosService;
+    @Autowired
+    private UsuarioService usuarioService;
+    @Autowired
+    private CompraService compraService;
 
     @GetMapping("")
     public String PagPrincipal(Model model) {
@@ -32,18 +32,21 @@ private CompraService compraService;
         model.addAttribute("libros", libros);
         return "admin/PagPrincipal";
     }
+
     @GetMapping("/usuarios")
-    public String Usuarios(Model model){
+    public String Usuarios(Model model) {
         model.addAttribute("usuarios", usuarioService.getAll());
         return "admin/usuarios";
     }
+
     @GetMapping("/compras")
-    public String Compras(Model model){
-       model.addAttribute("compras", compraService.getAll());
+    public String Compras(Model model) {
+        model.addAttribute("compras", compraService.getAll());
         return "admin/Compras";
     }
+
     @GetMapping("/detalles/{id}")
-    public String Detalles(Model model, @PathVariable Long id){
+    public String Detalles(Model model, @PathVariable Integer id) {
         Compra compra = compraService.findById(id).get();
         model.addAttribute("detalles", compra.getDetalleCompra());
         return "admin/DetallesCompras";
